+++ src/main/java/ir/sam/hearthstone/server/model/response/ResponseExecutor.java
@@ -36,7 +36,7 @@ public interface ResponseExecutor {
     void changeCardOnPassive(CardOverview cardOverview, int index);
 
     void setPlayDetail(List<PlayDetails.Event> events, String eventLog
-            , int[] mana, long time);
+            , int[] mana, double time);
 
     default void doLogout() {
     }
